Adjust form submission to work with client side apiClient
