+++ frontend/src/domain/User.ts
@@ -1,0 +1,10 @@
+type Roles = Array<string>
+
+export type User = {
+  id: number
+  username: string
+  email: string
+  firstName: string
+  lastName: string
+  roles: Roles
+}
